refactor: 인증 관련 설정 auth 패키지로 이동
