+++ backend/src/main/java/com/yagubogu/auth/token/JwtProvider.java
@@ -8,8 +8,8 @@ import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.yagubogu.auth.dto.MemberClaims;
-import com.yagubogu.global.config.JwtProperties;
-import com.yagubogu.global.config.JwtProperties.TokenProperties;
+import com.yagubogu.auth.config.JwtProperties;
+import com.yagubogu.auth.config.JwtProperties.TokenProperties;
 import com.yagubogu.global.exception.UnAuthorizedException;
 import com.yagubogu.member.domain.Role;
 import java.time.LocalDateTime;
