+++ backend/src/test/java/com/yagubogu/auth/token/JwtProviderTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.yagubogu.auth.dto.MemberClaims;
-import com.yagubogu.global.config.JwtProperties;
-import com.yagubogu.global.config.JwtProperties.TokenProperties;
+import com.yagubogu.auth.config.JwtProperties;
+import com.yagubogu.auth.config.JwtProperties.TokenProperties;
 import com.yagubogu.global.exception.UnAuthorizedException;
 import com.yagubogu.member.domain.Role;
 import org.junit.jupiter.api.BeforeEach;
